requirements.txt
astroid==2.5.2
autopep8==1.5.7
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
google-api-core==1.30.0
google-auth==1.31.0
google-cloud-core==1.7.0
google-cloud-speech==2.4.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
libcst==0.3.19
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.9
proto-plus==1.18.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pylint==2.7.2
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.5
wincertstore==0.2
wrapt==1.12.1
